Update README.md
This is unfortunately not true as we know. All subnets are in the same directory in the operating system. And as they always validate also with the primary (P) network, this is a MUST condition.  If something goes wrong in the subnet, you have to clear the whole database.  There is no way to rest only the subnet and restart it from the beginning.